Fix energy bars not rendering properly when theres a lot of power, closes https://github.com/BlakeBr0/ExtendedCrafting/issues/18

+++ src/main/java/com/blakebr0/extendedcrafting/client/gui/GuiCompressor.java
@@ -40,7 +40,7 @@ public class GuiCompressor extends GuiContainer {
 	private int getEnergyBarScaled(int pixels) {
 		int i = this.tile.getEnergy().getEnergyStored();
 		int j = this.tile.getEnergy().getMaxEnergyStored();
-		return j != 0 && i != 0 ? i * pixels / j : 0;
+		return (int) (j != 0 && i != 0 ? (long) i * pixels / j : 0);
 	}
 
 	private int getMaterialBarScaled(int pixels) {
