+++ src/main/java/com/blakebr0/extendedcrafting/client/gui/GuiCraftingCore.java
@@ -31,7 +31,7 @@ public class GuiCraftingCore extends GuiContainer {
 	private int getEnergyBarScaled(int pixels) {
 		int i = this.tile.getEnergy().getEnergyStored();
 		int j = this.tile.getEnergy().getMaxEnergyStored();
-		return j != 0 && i != 0 ? i * pixels / j : 0;
+		return (int) (j != 0 && i != 0 ? i * (long) pixels / j : 0);
 	}
 
 	private int getProgressBarScaled(int pixels) {
